+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.0.1
+
+* Reverse function added.
+* Return function added.
+* BIN Info function added.
+* Installment function added.
+* Point Inquiry function added.
+* Amounts are changed to double.
+* SDK min supported version decreased to 3.1.5
+
 ## 1.0.0
 
-* TODO: Initial release.
+* Initial release.